[Refactor] 주문 생성 시 사용하는 where 절을 메서드로 재구성
